+++ src/GodelTech.Microservices.Core/Mvc/RequestResponseLogging/RequestResponseLoggingMiddleware.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Net;
 using System.Text.Json;
 using System.Threading.Tasks;
-using GodelTech.Microservices.Core.Timer;
+using GodelTech.Microservices.Core.Utilities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Http.Features;
