+++ test/GodelTech.Microservices.Core.Tests/Mvc/RequestResponseLogging/RequestResponseLoggingMiddlewareTests.cs
@@ -7,7 +7,7 @@ using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
 using GodelTech.Microservices.Core.Mvc.RequestResponseLogging;
-using GodelTech.Microservices.Core.Timer;
+using GodelTech.Microservices.Core.Utilities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Http.Features;
